Fix crash on excessively wide menu
If text on a menu causes the menu to be wider than the screen, the
menu's x coordinate would be computed as a negative value.  Passing this
to the bitmap code would silently reinterpret it as a uint16_t, which
was then far above the maximum width of the parent canvas, triggering an
exception.

Fix this by clamping the x coordinate not to become negative.

Reported-by: WilliamWeberBerrutti <https://github.com/dxx-rebirth/dxx-rebirth/issues/736>

+++ common/main/newmenu.h
@@ -419,7 +419,7 @@ struct listbox_layout
 	}
 	void create_structure();
 	unsigned items_on_screen;
-	int box_x, box_y;
+	uint16_t box_x, box_y;
 	int box_w, height, title_height;
 	int citem, first_item;
 	unsigned nitems;
